+++ aquamarine/src/plumber.rs
@@ -475,7 +475,8 @@ mod tests {
         fn memory_stats(&self) -> AVMMemoryStats {
             AVMMemoryStats {
                 memory_size: 0,
-                max_memory_size: None,
+                total_memory_limit: None,
+                allocation_rejects: None,
             }
         }
     }