+++ crates/spell-service-api/src/lib.rs
@@ -324,7 +324,7 @@ mod tests {
 
         let workers = Arc::new(workers);
 
-        let max_heap_size = server_config::default_module_max_heap_size();
+        let service_memory_limit = server_config::default_service_memory_limit();
         let config = ServicesConfig::new(
             local_pid,
             base_dir,
@@ -332,8 +332,7 @@ mod tests {
             HashMap::new(),
             management_pid,
             root_key_pair.get_peer_id(),
-            max_heap_size,
-            None,
+            Some(service_memory_limit),
             Default::default(),
         )
         .unwrap();
@@ -342,8 +341,6 @@ mod tests {
             &config.modules_dir,
             &config.blueprint_dir,
             &config.particles_vault_dir,
-            max_heap_size,
-            None,
             Default::default(),
         );