+++ particle-services/src/app_services.rs
@@ -820,17 +820,18 @@ impl ParticleAppServices {
         let lock = service.service.lock();
         let stats = lock.module_memory_stats();
         let stats = stats
-            .0
+            .modules
             .into_iter()
             .map(|stat| {
                 json!({
                     "name": stat.name,
                     "memory_size_bytes": stat.memory_size,
-                    "max_memory_size_bytes": stat.max_memory_size
                 })
             })
             .collect();
 
+        // TODO: report service memory limit
+        // TODO: report allocation rejects (cleared after each call, optional value but always Some on wasmtime)
         Ok(stats)
     }
 
@@ -973,14 +974,15 @@ impl ParticleAppServices {
             .iter_mut()
             .for_each(|module| self.vault.inject_vault(module));
 
-        if let Some(metrics) = self.metrics.as_ref() {
-            metrics.observe_service_config(self.config.max_heap_size.as_u64(), &modules_config);
-        }
-
         let app_config = AppServiceConfig {
             service_working_dir: self.config.workdir.join(&service_id),
             service_base_dir: self.config.workdir.clone(),
             marine_config: MarineConfig {
+                // TODO: add an option to set individual per-service limit
+                total_memory_limit: self
+                    .config
+                    .default_service_memory_limit
+                    .map(|bytes| bytes.as_u64()),
                 modules_dir: Some(self.config.modules_dir.clone()),
                 modules_config,
                 default_modules_config: None,
@@ -1065,7 +1067,7 @@ mod tests {
         let vault_dir = base_dir.join("..").join("vault");
         let keypairs_dir = base_dir.join("..").join("keypairs");
         let workers_dir = base_dir.join("..").join("workers");
-        let max_heap_size = server_config::default_module_max_heap_size();
+        let service_memory_limit = server_config::default_service_memory_limit();
         let key_storage = KeyStorage::from_path(keypairs_dir.clone(), root_keypair.clone().into())
             .await
             .expect("Could not load key storage");
@@ -1094,8 +1096,7 @@ mod tests {
             HashMap::new(),
             management_pid,
             root_key_pair.get_peer_id(),
-            max_heap_size,
-            None,
+            Some(service_memory_limit),
             Default::default(),
         )
         .unwrap();
@@ -1104,8 +1105,6 @@ mod tests {
             &config.modules_dir,
             &config.blueprint_dir,
             &config.particles_vault_dir,
-            max_heap_size,
-            None,
             Default::default(),
         );
 
@@ -1193,8 +1192,6 @@ mod tests {
             file_name: None,
             load_from: None,
             config: TomlMarineModuleConfig {
-                mem_pages_count: None,
-                max_heap_size: None,
                 logger_enabled: None,
                 wasi: None,
                 mounted_binaries: None,
@@ -1236,8 +1233,6 @@ mod tests {
             file_name: None,
             load_from: None,
             config: TomlMarineModuleConfig {
-                mem_pages_count: None,
-                max_heap_size: None,
                 logger_enabled: None,
                 wasi: None,
                 mounted_binaries: None,